stage effort optimizer (still gives some weirdly large cells)

+++ src/main/java/de/uni_potsdam/hpi/asg/drivestrength/aggregatedcells/AggregatedCell.java
@@ -103,7 +103,7 @@ public class AggregatedCell {
         return sum / this.delayParameterTriples.size();
     }
     
-    public double getStageCountForPin(String pinName) {
+    public int getStageCountForPin(String pinName) {
     	return this.delayParameterTriples.get(pinName).getStageCount();
     }
